Move stories out of src
Otherwise stories are bundled during build.
