+++ .storybook/main.js
@@ -1,10 +1,10 @@
 module.exports = {
   "stories": [
-    "../src/**/*.stories.mdx",
-    "../src/**/*.stories.@(js|jsx|ts|tsx)"
+    "../storybook/**/*.stories.mdx",
+    "../storybook/**/*.stories.@(js|jsx|ts|tsx)"
   ],
   "addons": [
     "@storybook/addon-links",
     "@storybook/addon-essentials"
   ]
-}
+};
