+++ .storybook/preview.js
@@ -10,4 +10,4 @@ export const parameters = {
       date: /Date$/,
     },
   },
-}
+};
